jwt authentication fixes

+++ src/components/RequireAuth.tsx
@@ -1,40 +1,46 @@
-import React, {useEffect, useState} from 'react';
-import {useLocation} from "react-router-dom";
+import React, {useEffect} from 'react';
+import {Outlet, useNavigate} from "react-router-dom";
 import UserService from "../api/UserService";
-import {Backdrop, CircularProgress} from "@mui/material";
-import {useAppDispatch} from "../store";
-import {setUser} from "../store/userSlice";
+import {useAppDispatch, useAppSelector} from "../store";
+import {resetUser, selectToken, setUser} from "../store/userSlice";
 
 const RequireAuth = () => {
-    const location = useLocation();
-    const [isLoading, setIsLoading] = useState(false);
     const dispatch = useAppDispatch();
+    const navigate = useNavigate();
+    const token = useAppSelector(selectToken);
+
 
     useEffect(() => {
-        setIsLoading(true);
-        UserService.checkAuth()
-            .then(result => {
-                const { token } = result.data;
 
-                if (localStorage.getItem("token")) {
-                    localStorage.setItem("token", token);
-                } else {
-                    sessionStorage.setItem("token", token);
-                }
-                dispatch(setUser(result.data));
-            })
-            .catch(reason => {
-                console.log(reason);
-                localStorage.clear();
-                sessionStorage.clear();
-                dispatch(setUser({token: "", name: "", role: ""}));
-            })
-            .finally(() => setIsLoading(false))
-    }, [location])
+        if (!token && !localStorage.getItem("token") && !localStorage.getItem("token")) {
+            navigate("/login");
+            return;
+        }
 
-    return isLoading
-        ? <Backdrop open={isLoading}><CircularProgress/></Backdrop>
-        : null
+        if (!token) {
+            UserService.checkAuth()
+                .then(result => {
+                    const { token: newToken } = result.data;
+
+                    if (localStorage.getItem("token")) {
+                        localStorage.setItem("token", newToken);
+                    } else {
+                        sessionStorage.setItem("token", newToken);
+                    }
+
+                    dispatch(setUser(result.data));
+                })
+                .catch(reason => {
+                    console.log(reason); // dev-log
+                    localStorage.clear();
+                    sessionStorage.clear();
+                    dispatch(resetUser());
+                    navigate("/login");
+                })
+        }
+    }, [])
+
+    return <Outlet/>;
 };
 
 export default RequireAuth;
